feat: More scalars (#914)
- [x] basic types
- [x] date/time/duration types
- [x] interval types
- [x] struct types
- [x] decimal types
- [ ] ~map types~

+++ scalar/list_test.go
@@ -12,16 +12,16 @@ func TestListSet(t *testing.T) {
 		result List
 	}{
 		{source: []int{1, 2}, result: List{Value: []Scalar{
-			&Int64{Value: 1, Valid: true},
-			&Int64{Value: 2, Valid: true},
+			&Int{Value: 1, Valid: true},
+			&Int{Value: 2, Valid: true},
 		}, Valid: true, Type: arrow.ListOf(arrow.PrimitiveTypes.Int64)}},
 		{source: &List{Value: []Scalar{
-			&Int64{Value: 1, Valid: true},
-			&Int64{Value: 2, Valid: true},
+			&Int{Value: 1, Valid: true},
+			&Int{Value: 2, Valid: true},
 		}, Valid: true, Type: arrow.ListOf(arrow.PrimitiveTypes.Int64)},
 			result: List{Value: []Scalar{
-				&Int64{Value: 1, Valid: true},
-				&Int64{Value: 2, Valid: true},
+				&Int{Value: 1, Valid: true},
+				&Int{Value: 2, Valid: true},
 			}, Valid: true, Type: arrow.ListOf(arrow.PrimitiveTypes.Int64)}},
 	}
